Add a utility for storing different shader programs for Aquarium
Aquarium has a lot of different rendering settings that use different
shader programs. This patch adds a class that makes managing these
programs easier instead of requiring a separate line of code for
each combination of settings.

The programs are now compiled on demand instead of all being compiled
when loading.